+++ snapsplit/src/app/trip/[tripId]/budget/expense-add/page.tsx
@@ -1,0 +1,1 @@
+export { default } from "@/features/trip/[tripId]/budget/expense/add/AddExpensePage";
